Creates a services module.
The angular notation of [] in the module creates a new module, based
on the galleryServices file it imports.

The components module has a fancy regex for scooping up files in the
same folder -- trying to make this work with the hard-coded links to
files to make sure it makes sense.