Creates a services module which contains a gallery factory
Instead of creating factories in the index.js script (using the
module.factory('name', [importedFunction])), I create a module, and
register a factory in it. This follows the Angular docs and most blog
styling much more closely than the regex script Marty uses (which is
fantastically unclear as to what is happening).

Possibly redundant modularization with the factory creation - but
it does pull in the factory correctly.