Refactors the services to be created in the index, instead of the service.
Was creating the factory as an Angular module and then registering the
module with the services module. This creates an unnecessary layer
of modules. With the refactor, it creates one services module and
then imports the functionality of the galleryService service in
using the es6 importing system, instead of Angular's modules system.