Updates image-app component to have add/remove. Removes name-spacing to app instead of using $ctrl. Passes component tests.
Added two more of the CRUD routes to the Angular front-end -- it
can now call the galleryService to call the backend for both an
POST and a DELETE route.

The controller had been name-spaced as app, instead of using the
default $ctrl, but David Goodwin mention that that makes components
less reuseable because the can start to have name-space collisions.
Which is funny, because the point of name spacing it was to avoid
collisions. tl;dr namespace is good if you don't use component
architecture.

The tests mock the galleryService to test that the main app component
works. Passes a get all to load and an add new.