Creates a component to add new images using galleryService
The new component has three text imput fields for title, url, desc,
which make an object to be sent to the backend. Can see the object
shows up using Postman to GET. Moves the add() function to this
component, taking it off of the image-app.

Has some basic styling -- needs to overwrite some of the default
button styling though.