Adds the add-image to the index page, removes add from image-app
There was an unused add() function in the image-app, moved it's
functionality into the add-image component instead. The 'push to
array' functionality wasn't needed because the GET will grab all
the images for me instead. Not sure this is how this should work.

Remove is still on the image-app, it might belong elsewhere.