Adds Angular's UI-Router to app
The router is added to the app in the app.js as an app.use on the
included routes script. The routes.js uses two helpers provided by
ui-router, the $stateProvider and $urlRouterProvider. These let
you define 'states' which should be loaded based on user interaction
with #! front-side routes. The routes can load a component or a
template.

I have a basic landing template which is loaded on hitting /
it's mostly poorly styled instructions.

Not testable without using protractor.js