Transitions page to have about links in sidebar
The about "page" is no longer a navigatable page, as indicated by
the "abstract: true" property. Instead, it has two child views
which will display in the main body when they're clicked.

The ui-view on the index page is where the about page hangs it's
ui-view for it's children to populate. You need both to get the
templates to load.