Adds parent 'about' component and children about app and about dev
The About component is an 'abstract' component, meaning it's not
able to be loaded, but it has two sub views for the about pages
which can be loaded.

The goal is to move these into sub/nested named views on the parent
component. Getting it working in a more basic configuration before
switching it over.