Fixes double styling on buttons - buttons are correct height again.
Adding new css for the header, left nav, and main body was shadowing
the class for the buttons -- both were using "main", so the buttons
were getting height properties from the main body.

Changes the names of the styling classes for the layout to be more
descriptive. Also removes the margin property from the button mixin
since it doesn't make sense given the new layout.