Updates mongoose models to have image and album
The old 'gallery' model was misnamed to begin with, since it was a
model for the images, not the gallery. It's been renamed and given
an external field to link to the album model.

Album model just has a name and description, nothing about the
images themselves. Routes and tests need to be updated.