Updates routes to use albums and images
If we're going to have albums as a view on the front end, it makes
more sense for the Get-all route to return the list of albums,
rather than all images.

The get with a name for album does a find by album name and returns
that album's data, but not any of it's images (yet).

Added a post route for adding a new album - works in Postman.