Updates template and add-image controller to add images to an album
The ui-router deals with getting the album data to populate the albums
dropdown menu *before* the component is created and put on the page.
The ng-options uses album as album.title to set the label and text
value of the options tag, and the track by albums._id sets the value
of the options tag.

Inside the controller, the add() function scrapes the data out of
the fields and builds the image object mongoose is expecting. Add
calls the galleryService's add function, which calls the backend
API's add image route. Successfully adds images to an album
(checked via Postman -- not tested).