Creates component for albums - js controller, html, and scss
The albums template repeats over the albums array and puts spans
with the title and links to the sub routes on the page.

Binds the albums array, display type, and db id in.

Controller doesn't currently do much, at some point should probably
have the code for adding a new album.