Updates routes to resolve images in parent, and correctly link to components.
The links to the components were in kabob case instead of camelCase,
so nothing was being linked correctly. This changes the components
to point to the right things.

The "images" were supposed to be resolved in the album controller
in an $onInit function, but they weren't being properly passed to
the child states. This moves the image gathering from the controller
to the router's $resolve function. Now the child states have access
to the images array via the bindings.