Moves old image-app into just buttons to toggle image views
This component had been the heart of the old app, but all it really
does is toggle between views. Instead, it now is a component which
is displayed at the top of the 'body' section of the page.

Changed from using a view==='detail' type click to using the ui-
router's helper ui-sref to change what is being put into the
parent's <ui-view> tag.

The click _should_ be sending display information for the url to
have params, but it's not currently working. Buttons toggle between
views just fine.