Removes unused stylesheet, $onInit function from controller
The $onInit function was intended to act as a hook to grab the
images from the galleryService when the Album controller was
instantiated. Didn't work for getting the images array to the
child states through the bindings.

Moved the images to the router in a $resolve block instead and it
passes the data to child states as intended.

Added the galleryService's remove function to try and add a remove
button to the thumbnail view.