Removes galleryService from component as it wasn't doing anything
I was under the impression that you could put the functionality on
the parent state and have the child states get access through it
via explicit bindings. This wouldn't work in any incantation for
the delete. Moved it out of the parent into the child.

Delete works, but doesn't dynamically update the page, so it doesn't
seem to work.